Upgrade to go1.10 and regenerate
Go 1.10 makes the move to the new context automatically on `go fix` so

Fixes: #326
