+++ examples/server/a_bit_of_everything.go
@@ -5,6 +5,7 @@ import (
 	"io"
 	"sync"
 
+	"context"
 	"github.com/golang/glog"
 	"github.com/golang/protobuf/proto"
 	"github.com/golang/protobuf/ptypes/duration"
@@ -13,7 +14,6 @@ import (
 	sub "github.com/grpc-ecosystem/grpc-gateway/examples/sub"
 	sub2 "github.com/grpc-ecosystem/grpc-gateway/examples/sub2"
 	"github.com/rogpeppe/fastuuid"
-	"golang.org/x/net/context"
 	"google.golang.org/genproto/googleapis/rpc/errdetails"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
