+++ examples/server/echo.go
@@ -1,9 +1,9 @@
 package server
 
 import (
+	"context"
 	"github.com/golang/glog"
 	examples "github.com/grpc-ecosystem/grpc-gateway/examples/examplepb"
-	"golang.org/x/net/context"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/metadata"
 )
