+++ runtime/context_test.go
@@ -6,8 +6,8 @@ import (
 	"testing"
 	"time"
 
+	"context"
 	"github.com/grpc-ecosystem/grpc-gateway/runtime"
-	"golang.org/x/net/context"
 	"google.golang.org/grpc/metadata"
 )
