+++ runtime/errors_test.go
@@ -8,8 +8,8 @@ import (
 	"strings"
 	"testing"
 
+	"context"
 	"github.com/grpc-ecosystem/grpc-gateway/runtime"
-	"golang.org/x/net/context"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 )
