+++ runtime/mux.go
@@ -6,8 +6,8 @@ import (
 	"net/textproto"
 	"strings"
 
+	"context"
 	"github.com/golang/protobuf/proto"
-	"golang.org/x/net/context"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/metadata"
 	"google.golang.org/grpc/status"
