+++ runtime/proto_errors.go
@@ -4,7 +4,7 @@ import (
 	"io"
 	"net/http"
 
-	"golang.org/x/net/context"
+	"context"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/grpclog"
 	"google.golang.org/grpc/status"
